capture username after login, moved upload fn to helper
