+++ src/features/Auth/authSlice.js
@@ -4,6 +4,7 @@ const initialState = {
   displayName: "",
   uid: "",
   email: "",
+  username: "",
   photoURL: "",
   isLoggedIn: false,
   error: "",
@@ -14,12 +15,13 @@ const authSlice = createSlice({
   initialState,
   reducers: {
     loginAction: (state, action) => {
-      const { displayName, uid, email, photoURL } = action.payload;
+      const { displayName, uid, email, photoURL, username } = action.payload;
       state.displayName = displayName;
       state.uid = uid;
       state.email = email;
       state.photoURL = photoURL;
       state.isLoggedIn = true;
+      state.username = username;
     },
     logoutUser: (state) => {
       state.displayName = "";
@@ -27,6 +29,7 @@ const authSlice = createSlice({
       state.email = "";
       state.photoURL = "";
       state.uid = "";
+      state.username = "";
     },
   },
 });
