+++ src/index.css
@@ -1,3 +1,23 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+::-webkit-scrollbar {
+  width: 12px;
+  height: 12px;
+}
+
+::-webkit-scrollbar-track {
+  border-radius: 100vh;
+  background: hsl(335, 77%, 98%);
+}
+
+::-webkit-scrollbar-thumb {
+  background: hsl(335, 77%, 70%);
+  border-radius: 100vh;
+  border: 3px solid #f6f7ed;
+}
+
+::-webkit-scrollbar-thumb:hover {
+  background: hsl(335, 77%, 60%);
+}
